Add a complex test for local variable scopes
It ensures that changes regarding the scope of local variables are accounted for by the similarity checking.

This test has shown that there are issues regarding the definition of similarity within CIPM. Currently disabled for the sake of not causing issues regarding Git or the building process.

Note: This test reproduces an issue, which emerged while using CIPM on TeaStore (propagation from v1.0.0 to v1.1.0). The said issue is about the comparison of the container of LocalVariableStatement instances, which are not directly contained by their intended container (example: TryBlock (intended container) -> Block (allows TryBlock to contain statements) -> LocalVariableStatement).